BE-58: Resolve conflicts

+++ src/main/java/az/gdg/msarticle/repository/ArticleRepository.java
@@ -1,6 +1,8 @@
 package az.gdg.msarticle.repository;
 
 import az.gdg.msarticle.model.entity.ArticleEntity;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -13,7 +15,7 @@ public interface ArticleRepository extends MongoRepository<ArticleEntity, String
 
     List<ArticleEntity> getArticleEntitiesByUserIdAndIsDraft(Integer userId, boolean isDraft);
 
-    Page<ArticleEntity> getArticleEntitiesByUserId(Integer userId, Pageable pageable);
+    Page<ArticleEntity> getArticleEntitiesByUserId(Long userId, Pageable pageable);
 
-    Page<ArticleEntity> getArticleEntitiesByUserIdAndIsDraftFalseAndIsApprovedTrue(Integer userId, Pageable pageable);
+    Page<ArticleEntity> getArticleEntitiesByUserIdAndIsDraftFalseAndIsApprovedTrue(Long userId, Pageable pageable);
 }
