+++ src/main/java/az/gdg/msarticle/service/ArticleService.java
@@ -1,6 +1,7 @@
 package az.gdg.msarticle.service;
 
 import az.gdg.msarticle.model.dto.ArticleDTO;
+import az.gdg.msarticle.model.dto.UserArticleDTO;
 
 public interface ArticleService {
 
@@ -12,5 +13,5 @@ public interface ArticleService {
 
     String publishArticle(String articleId);
 
-    UserArticleDTO getArticlesByUserId(Integer userId, int page);
+    UserArticleDTO getArticlesByUserId(Long userId, int page);
 }
